requirements.txt
argparse
pyyaml
python-dotenv
pytest-playwright
url-normalize